slightly refactor

+++ lib/pervasives.ml
@@ -6,6 +6,6 @@ external sin : float -> float = "sin"
 external atan : float -> float = "atan"
 
 external float_of_int : int -> float = "float_of_int"
-external int_of_float : int -> float = "int_of_float"
+external int_of_float : float -> int = "int_of_float"
 
 external print_int : int -> unit = "print_int"
